image: parse image to offsets rather than addresses
When parsing images, we currently store pointers to interesting parts of
the image in pointers to within the image buffer. However, this causes
problems when reallocating the buffer (eg, when we append the
signature), as the pointers may no longer be valid. So, we have to
re-parse.

Rather than keeping pointers to the image, this change uses offsets
instead. We parse to (unsigned int) offsets in struct image, and
add a few accessors to simplify reading from structures in the image
data.

Signed-off-by: Jeremy Kerr <jeremy.kerr@canonical.com>